fix: resolve multiple compilation errors across modules
- ProductRepository: add countByInventoryStatusAndIsActiveTrue, reserveStockAndIsActiveTrue, releaseReservedStockAndIsActiveTrue
- OrderRepository: add countByPaymentStatusAndIsActiveTrue, calculateTotalRevenue, findBySellerId methods
- AuthService: add cleanupExpiredSessions method
- AuthResponse: add tokenType field
- CartServiceImpl: fix ResourceNotFoundException usage
- CategoryMapper: add toEntityFromRequest alias method
- ProductFilterRequest: change Float to BigDecimal for rating fields

+++ backend/src/main/java/ecommerce/modules/product/dto/ProductFilterRequest.java
@@ -18,8 +18,8 @@ public class ProductFilterRequest {
     private String categorySlug;
     private BigDecimal minPrice;
     private BigDecimal maxPrice;
-    private Float minRating;
-    private Float maxRating;
+    private BigDecimal minRating;
+    private BigDecimal maxRating;
     private String search;
     private String keyword;
     private String name;
